LoginBackend/requirements.txt
fastapi
azure-data-tables
python-dotenv
pydantic
uvicorn
bcrypt
